Fixing bugs on removing file async and inclued action to check if file exists before delete it

+++ services/file.service.js
@@ -23,9 +23,14 @@ const fileService = {
 
   removeFilesAsync: async (files) => {
     for await (const file of files) {
-      fs.unlink(path.join(config.uploadDir, "/", file), (err) => {
-        if (err) console.log(err);
-      });
+      let absolutePath = path.join(config.uploadDir, '/', file);
+      fs.exists(absolutePath, (exists) => {
+        if (exists) {
+          fs.unlink(path.join(config.uploadDir, '/', file), (err) => {
+            if (err) console.log(err);
+          });
+        }       
+      });    
     }
   }
 };
